Reworked FileUtil
